+++ vpin-studio-commons/src/main/java/de/mephisto/vpin/commons/fx/OverlayWindowFX.java
@@ -168,6 +168,12 @@ public class OverlayWindowFX extends Application {
     });
   }
 
+  public void exitPauseMenu() {
+    Platform.runLater(() -> {
+      PauseMenu.exit();
+    });
+  }
+
   public void showHighscoreCard(CardSettings cardSettings, PinUPPlayerDisplay display, File file) {
     try {
       int notificationTime = cardSettings.getNotificationTime();