+++ vpin-studio-commons/src/main/java/module-info.java
@@ -22,6 +22,7 @@ open module de.mephisto.vpin.commons {
   requires de.mephisto.vpin.restclient;
   requires jnativehook;
   requires java.logging;
+  requires de.mephisto.vpin.connectors.vps;
 
   exports de.mephisto.vpin.commons;
   exports de.mephisto.vpin.commons.fx;