+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/preferences/PauseMenuSettings.java
@@ -1,0 +1,24 @@
+package de.mephisto.vpin.restclient.preferences;
+
+import de.mephisto.vpin.restclient.JsonSettings;
+
+public class PauseMenuSettings extends JsonSettings {
+  private String key;
+  private boolean useOverlayKey;
+
+  public String getKey() {
+    return key;
+  }
+
+  public void setKey(String key) {
+    this.key = key;
+  }
+
+  public boolean isUseOverlayKey() {
+    return useOverlayKey;
+  }
+
+  public void setUseOverlayKey(boolean useOverlayKey) {
+    this.useOverlayKey = useOverlayKey;
+  }
+}