+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/SplashScreenController.java
@@ -16,14 +16,20 @@ public class SplashScreenController implements Initializable {
   @FXML
   private Label versionLabel;
 
+  @FXML
+  private Label factLabel;
+
+
+
   @Override
   public void initialize(URL location, ResourceBundle resources) {
     versionLabel.setText("Version " + Studio.getVersion());
-
+    factLabel.setText("");
     try {
       List<String> facts = IOUtils.readLines(SplashScreenController.class.getResourceAsStream("facts.txt"));
       int i = ThreadLocalRandom.current().nextInt(0, facts.size() - 1);
       String fact = facts.get(i);
+      factLabel.setText("\"" + fact + "\"");
     } catch (IOException e) {
       throw new RuntimeException(e);
     }