+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/Studio.java
@@ -179,6 +179,7 @@ public class Studio extends Application {
 
         client.setErrorHandler(errorHandler);
 
+        LOG.info("Init finished, showing Studio stage.");
         stage.show();
         splash.hide();
       });