+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 - Removed "ID" column from the tables overview (no need to display this technical value).
 - Added column "Date Added" to table overview. Note that this value may be empty because previous version of the Studio did not set this value when new tables have been uploaded.
 - Improved "Players" section: Added default background image for score items.
+- Added fancy splash screen.
 
 ### Bugfixes